+++ examples/flask_example.py
@@ -5,15 +5,15 @@ from typing import Any
 from flask import Flask
 
 import puff
+from puff import wrap_async
 import puff.postgres
+from puff.redis import global_redis
 import redis
 
 SAMPLE = 1000
 
 app = Flask(__name__)
 r = redis.Redis(host="localhost")
-
-puff_redis = puff.global_redis()
 
 
 @app.route("/pg/")
@@ -50,11 +50,8 @@ def hello_world():
     return b"".join(gets)
 
 
-from puff import wrap_async
-
-
 def redis_get(q):
-    return wrap_async(lambda rr: puff_redis.get(rr, q), join=False)
+    return global_redis.get(q)
 
 
 def redis_get_many(q, l):
@@ -64,41 +61,4 @@ def redis_get_many(q, l):
 
 
 def redis_set(k, v):
-    return wrap_async(lambda rr: puff_redis.set(rr, k, v), join=False)
-
-
-# def app(environ, start_response):
-#     start_response("200 OK", [])
-#     redis_set("blam", "ok")
-#     gets = []
-#     for _ in range(SAMPLE):
-#         gets.append(redis_get("blam"))
-#     return [b"".join(gets)]
-#
-#
-# def do_work(q, return_result):
-#     return_result(q)
-#
-#
-# def redis_get(q):
-#     return wrap_async(lambda rr: do_work(q, rr))
-#
-#
-# def my_job():
-#     r = redis_get("blam")
-#     print(f"Got {r} from redis.")
-#     r = redis_get("blam2")
-#     print(f"Got {r} from redis again.")
-#
-#
-# t = start_event_loop()
-#
-#
-# def returner(r):
-#     print(f"Got {r} as a value.")
-#
-#
-# t.spawn(my_job, [], {}, returner)
-# t.spawn(my_job, [], {}, returner)
-# t.spawn(my_job, [], {}, returner)
-# t.spawn(my_job, [], {}, returner)
+    return global_redis.set(k, v)