+++ examples/python_tests/test_pubsub.py
@@ -1,14 +1,12 @@
 import asyncio
 
-from puff import global_pubsub, spawn, sleep_ms
+from puff import spawn, sleep_ms
+from puff.pubsub import global_pubsub
 from asgiref.sync import async_to_sync, sync_to_async
 
 
-pubsub = global_pubsub()
-
-
 async def wait_for_message():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     await conn.subscribe("test-chan")
     msg = await conn.receive()
     data = msg.json()
@@ -17,14 +15,14 @@ async def wait_for_message():
 
 async def publish_and_wait():
     task = asyncio.create_task(wait_for_message())
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     await sleep_ms(100)
     await conn.publish_json("test-chan", {"result": 1})
     assert await task == 1
 
 
 def wait_for_message_greenlet():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     conn.subscribe("test-chan")
     msg = conn.receive()
     data = msg.json()
@@ -33,7 +31,7 @@ def wait_for_message_greenlet():
 
 def publish_and_wait_greenlet():
     greenlet = spawn(wait_for_message_greenlet)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish_json("test-chan", {"result": 2})
     assert greenlet.join() == 2
@@ -41,14 +39,14 @@ def publish_and_wait_greenlet():
 
 async def publish_and_wait_sync_from_async():
     task = asyncio.create_task(sync_to_async(wait_for_message_greenlet)())
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     await asyncio.sleep(0.1)
     await conn.publish_json("test-chan", {"result": 3})
     assert await task == 3
 
 
 def wait_for_message_bytes():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     conn.subscribe("test-chan")
     msg = conn.receive()
     return msg.body
@@ -56,14 +54,14 @@ def wait_for_message_bytes():
 
 def test_publish_and_wait_bytes():
     greenlet = spawn(wait_for_message_bytes)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish_bytes("test-chan", b"123456")
     assert greenlet.join() == b"123456"
 
 
 def wait_for_message_string():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     conn.subscribe("test-chan")
     msg = conn.receive()
     return msg.text
@@ -71,14 +69,14 @@ def wait_for_message_string():
 
 def test_publish_and_wait_string():
     greenlet = spawn(wait_for_message_string)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish("test-chan", "123456")
     assert greenlet.join() == "123456"
 
 
 def wait_for_message_string_multi():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     conn.subscribe("test-chan")
     msg = conn.receive()
     msg2 = conn.receive()
@@ -87,7 +85,7 @@ def wait_for_message_string_multi():
 
 def test_publish_and_wait_string_multi():
     greenlet = spawn(wait_for_message_string_multi)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish("test-chan", "123456")
     conn.publish("test-chan", "789")
@@ -95,7 +93,7 @@ def test_publish_and_wait_string_multi():
 
 
 def wait_for_message_string_multi_channels():
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     conn.subscribe("test-chan-3")
     conn.subscribe("test-chan-2")
     conn.subscribe("test-chan-1")
@@ -107,7 +105,7 @@ def wait_for_message_string_multi_channels():
 
 def test_publish_and_wait_string_multi_channels():
     greenlet = spawn(wait_for_message_string_multi_channels)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish("test-chan-1", "123456")
     conn.publish("test-chan-2", "789")
@@ -119,7 +117,7 @@ def test_publish_and_wait_string_multi_channels_multi_wait():
     greenlet = spawn(wait_for_message_string_multi_channels)
     greenlet2 = spawn(wait_for_message_string_multi_channels)
     greenlet3 = spawn(wait_for_message_string_multi_channels)
-    conn = pubsub.connection()
+    conn = global_pubsub.connection()
     sleep_ms(100)
     conn.publish("test-chan-1", "123456")
     conn.publish("test-chan-2", "789")
@@ -127,6 +125,30 @@ def test_publish_and_wait_string_multi_channels_multi_wait():
     assert greenlet.join() == ("123456", "789", "abc")
     assert greenlet2.join() == ("123456", "789", "abc")
     assert greenlet3.join() == ("123456", "789", "abc")
+
+
+def wait_for_message_string_unsubscribe():
+    conn = global_pubsub.connection()
+    conn.subscribe("test-chan-3")
+    conn.subscribe("test-chan-2")
+    conn.subscribe("test-chan-1")
+    conn.unsubscribe("test-chan-2")
+    msg = conn.receive()
+    msg3 = conn.receive()
+    return msg.text, msg3.text
+
+
+def test_publish_and_wait_string_multi_channels_multi_wait():
+    greenlet = spawn(wait_for_message_string_unsubscribe)
+    greenlet2 = spawn(wait_for_message_string_multi_channels)
+    conn = global_pubsub.connection()
+    sleep_ms(100)
+    conn.publish("test-chan-1", "123456")
+    conn.publish("test-chan-2", "789")
+    conn.publish("test-chan-3", "abc")
+    assert greenlet.join() == ("123456", "abc")
+    assert greenlet2.join() == ("123456", "789", "abc")
+
 
 
 def test_pubsub_async():